requirements.txt
numpy
pandas
Pillow
scipy
matplotlib
tensorflow-cpu == 2.3.0
keras == 2.3.0